Add unit tests for packmol SCMSUITE-8674 SO--
